Combine heat/hvac into climate
This solves later issues with circular dependencies.
